Fixed remove_old_results script

+++ src/other/build_assets/remote_testing/remove_old_results.sh
@@ -7,15 +7,22 @@ DIR=$1
 if [ -z "$DIR" ]
 then
     echo "Directory containing test results required as a parameter"
+    exit 1
 fi
 
 echo "Removing old test results"
 
 cd $DIR
+if [ $? -gt 0 ]
+then
+    echo "Cannot access directory $DIR"
+    exit 1
+fi
+
 echo $(pwd)
 
-TO_REMOVE=`ls -1r | tail -n +4`
-echo "Will remove this old test results:"
+TO_REMOVE=`ls -1t | tail -n +4`
+echo "Will remove these old test results:"
 echo $TO_REMOVE
 
 rm -rf $TO_REMOVE
